+++ Swiftfin tvOS/Views/ItemView/Components/ActionButtonView/Components/ActionButton/ActionButton.swift
@@ -128,81 +128,3 @@ extension ItemView.ActionButton {
         self.content = content
     }
 }
-
-struct ActionButtonStyle: ButtonStyle {
-
-    // MARK: - Focus State
-
-    @Environment(\.isFocused)
-    private var isFocused
-
-    func makeBody(configuration: Configuration) -> some View {
-        GeometryReader { geometry in
-            configuration.label
-                .modifier(ActionButtonScaleModifier(
-                    size: geometry.size,
-                    expansion: getExpansionAmount(isPressed: configuration.isPressed),
-                    animationDuration: 0.15,
-                    isPressed: configuration.isPressed,
-                    isFocused: isFocused
-                ))
-        }
-    }
-
-    // MARK: - Expand a Fixed Pixel Amount
-
-    private func getExpansionAmount(isPressed: Bool) -> CGFloat {
-        if isPressed {
-            return 4
-        } else if isFocused {
-            return 16
-        } else {
-            return 0
-        }
-    }
-}
-
-// MARK: - Fixed-Size Scale Modifier
-
-/// Since these buttons are variable in size, a 10% increase is too much for the largest button size but too little for the smallest button
-/// This Modifier ensures a fixed pixel expansion/contraction is used to ensure a consistent animation
-struct ActionButtonScaleModifier: ViewModifier {
-
-    // MARK: - Button Properties
-
-    let size: CGSize
-
-    // MARK: - Expansion Properties
-
-    let expansion: CGFloat
-    let animationDuration: Double
-
-    // MARK: - Expansion Reason(s)
-
-    let isPressed: Bool
-    let isFocused: Bool
-
-    // MARK: - Body
-
-    func body(content: Content) -> some View {
-        content
-            .scaleEffect(calculateScaleFactor())
-            .animation(
-                .easeInOut(duration: animationDuration),
-                value: isPressed || isFocused
-            )
-    }
-
-    // MARK: - Calculate Percentage from Properties
-
-    private func calculateScaleFactor() -> CGFloat {
-        if expansion == 0 {
-            return 1.0
-        }
-        let baseSize = max(size.width, size.height)
-
-        guard baseSize > 0 else { return 1.0 }
-
-        return (baseSize + (2 * expansion)) / baseSize
-    }
-}
